+++ clients/python/graphsense/__init__.py
@@ -12,7 +12,7 @@
 """
 
 
-__version__ = "2.9.5"
+__version__ = "2.9.6"
 
 # Define package exports
 __all__ = [